+++ js/data.js
@@ -1,0 +1,46 @@
+export const TITLES = [
+  'Сдам квартиру',
+  'Сдам комнату',
+  'Сдам гараж',
+];
+export const PRICE_MIN = 10;
+export const PRISE_MAX = 1000;
+export const ROOMS_MIN = 1;
+export const ROOMS_MAX = 10;
+export const GUESTS_MIN = 1;
+export const GUESTS_MAX = 50;
+export const TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel',
+];
+export const CHECKIN_CHECKOUT_TIME = [
+  '12:00',
+  '13:00',
+  '14:00',
+];
+export const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner',
+];
+export const DESCRIPTIONS = [
+  'Вид на море',
+  'Вид во двор',
+  'Вид на дорогу',
+];
+export const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg',
+];
+export const LAT_MIN = 35.65;
+export const LAT_MAX = 35.7;
+export const LNG_MIN = 139.7;
+export const LNG_MAX = 139.8;
+export const COORDINATES_ACCURACY = 5;
